scoring/requirements.txt
spacy>=3.7
scikit-learn>=1.3
pdfminer.six>=20221105
openpyxl>=3.1
